test: init memory
