+++ src/tests/utils/console-tests.utils.ts
@@ -460,3 +460,20 @@ export const assertAssetServed = async ({
 
 	expect(decoder.decode(body as Uint8Array<ArrayBufferLike>)).toEqual(content);
 };
+
+export const updateRateConfig = async ({
+	actor
+}: {
+	actor: Actor<ConsoleActor_0_0_8 | ConsoleActor_0_0_14 | ConsoleActor>;
+}) => {
+	const { update_rate_config } = actor;
+
+	const config = {
+		max_tokens: 100n,
+		time_per_token_ns: 60n
+	};
+
+	await update_rate_config({ Satellite: null }, config);
+	await update_rate_config({ Orbiter: null }, config);
+	await update_rate_config({ MissionControl: null }, config);
+};
